Orbital: Correct Success Logic for Refund
According to Orbital documentation responses for all new orders contain ProcStatus, ApprovalStatus, RespCode. The success of transaction should be determined based on combination of these values: ProcStatus 0=Success, ApprovalStatus 0=Declined 1=Approved 2=Message/System Error, RespCode 00 Approved and some other values as per table in documentation. The case when RespCode is empty is not desctibed in documentation, but since empty RespCode is not listed in the table as Approved implementation based on the assumption that empty RespCode is failure. Remote tests for eCheck refunds are failing due to this case. The response has ApprovalStatus but empty RespCode.
ECS-1881

Unit:
125 tests, 726 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
69 tests, 315 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
97.1014% passed

Failed tests are due to issue described above